LPS-76715 Move
